pyproject.toml
[project]
name = "hard-negatives-ranking-datasets-maker"
version = "0.1.0"
description = "Mine hard negatives for retrieval training using ColBERT (PyLate)"
readme = "README.md"
requires-python = "==3.10"
dependencies = [
    "transformers",
    "sentence-transformers",
    "datasets==3.6.0",
    "flash_attn @ https://github.com/mjun0812/flash-attention-prebuild-wheels/releases/download/v0.4.11/flash_attn-2.8.3+cu128torch2.8-cp310-cp310-linux_x86_64.whl",
    "python-dotenv",
    "pyyaml",
    "torch==2.8.0",
    "evaluate",
    "scikit-learn",
    "accelerate>=0.26.0",
    "wandb",
    "mteb",
    "pylate>=1.3.3",
    "fast-plaid",
    "tabulate",
]
